Move kpoints into convergence component
