+++ src/aiidalab_qe/app/configuration/advanced/convergence/model.py
@@ -3,6 +3,10 @@ from __future__ import annotations
 import numpy as np
 import traitlets as tl
 
+from aiida import orm
+from aiida_quantumespresso.calculations.functions.create_kpoints_from_distance import (
+    create_kpoints_from_distance,
+)
 from aiida_quantumespresso.workflows.pw.base import PwBaseWorkChain
 from aiidalab_qe.common.mixins import HasInputStructure
 
@@ -13,7 +17,7 @@ class ConvergenceConfigurationSettingsModel(
     AdvancedCalculationSubSettingsModel,
     HasInputStructure,
 ):
-    title = "Convergence"
+    title = "Convergence/Accuracy"
     identifier = "convergence"
 
     dependencies = [
@@ -30,6 +34,8 @@ class ConvergenceConfigurationSettingsModel(
     etot_conv_thr_step = tl.Float(1e-5)
     forc_conv_thr = tl.Float(0.0)
     forc_conv_thr_step = tl.Float(1e-4)
+    kpoints_distance = tl.Float(0.0)
+    mesh_grid = tl.Unicode("")
     electron_maxstep = tl.Int(80)
     optimization_maxsteps = tl.Int(50)
     mixing_mode_options = tl.List(
@@ -44,10 +50,15 @@ class ConvergenceConfigurationSettingsModel(
     mixing_beta = tl.Float(0.4)
 
     def update(self, specific=""):
-        if specific == "structure":
-            self._update_help_message()
-        elif specific == "protocol":
-            self._update_thresholds()
+        with self.hold_trait_notifications():
+            if specific == "structure":
+                self._update_help_message()
+            elif specific == "protocol":
+                self._update_thresholds()
+            if not specific or specific != "mesh":
+                parameters = PwBaseWorkChain.get_protocol_inputs(self.protocol)
+                self._update_kpoints_distance(parameters)
+            self._update_kpoints_mesh()
 
     def reset(self):
         with self.hold_trait_notifications():
@@ -57,6 +68,7 @@ class ConvergenceConfigurationSettingsModel(
             self.etot_conv_thr_step = self._get_default("etot_conv_thr_step")
             self.forc_conv_thr = self._get_default("forc_conv_thr")
             self.forc_conv_thr_step = self._get_default("forc_conv_thr_step")
+            self.kpoints_distance = self._get_default("kpoints_distance")
             self.electron_maxstep = self._get_default("electron_maxstep")
             self.optimization_maxsteps = self._get_default("optimization_maxsteps")
             self.mixing_mode = self._get_default("mixing_mode")
@@ -107,6 +119,26 @@ class ConvergenceConfigurationSettingsModel(
         else:
             step = 0.1
         self._defaults[f"{attribute}_step"] = step
+
+    def _update_kpoints_distance(self, parameters):
+        kpoints_distance = parameters["kpoints_distance"] if self.has_pbc else 100.0
+        self._defaults["kpoints_distance"] = kpoints_distance
+        self.kpoints_distance = self._defaults["kpoints_distance"]
+
+    def _update_kpoints_mesh(self, _=None):
+        if not self.has_structure:
+            mesh_grid = ""
+        elif self.kpoints_distance > 0:
+            mesh = create_kpoints_from_distance.process_class._func(
+                self.input_structure,
+                orm.Float(self.kpoints_distance),
+                orm.Bool(False),
+            )
+            mesh_grid = f"Mesh {mesh.get_kpoints_mesh()[0]!s}"
+        else:
+            mesh_grid = "Please select a number higher than 0.0"
+        self._defaults["mesh_grid"] = mesh_grid
+        self.mesh_grid = self._get_default("mesh_grid")
 
     def _get_default(self, trait):
         return self._defaults.get(trait, self.traits()[trait].default_value)
